UI 调整，适配竖屏
